moves pairing into seperate module, enables setting non static passkey (is randomized after each conn attempt
