+++ src/ble.c
@@ -67,17 +67,13 @@
 #include "nrf_sdh_ble.h"
 #include "app_timer.h"
 
-//#include "nrf_ble_lesc.h"
-#include "fds.h"
-#include "peer_manager.h"
-#include "peer_manager_handler.h"
-#include "ble_conn_state.h"
-#include "nrf_ble_lesc.h"
-
 #include "bsp_btn_ble.h"
 #include "nrf_ble_qwr.h"
 #include "nrf_ble_gatt.h"
 #include "nrf_pwr_mgmt.h"
+
+#include "peer_manager_handler.h"
+#include "nrf_ble_lesc.h"
 
 #include "nrf_log.h"
 #include "nrf_log_ctrl.h"
@@ -86,14 +82,13 @@
 
 #include "ble_config.h"
 #include "service_if.h"
+#include "pairing.h"
 
 NRF_BLE_QWR_DEF(m_qwr);                                                             /**< Queued Writes structure.*/
 NRF_BLE_GATT_DEF(m_gatt);                                                           /**< GATT module instance. */
 BLE_ADVERTISING_DEF(m_advertising);                                                 /**< Advertising module instance. */
 
-static pm_peer_id_t m_peer_to_be_deleted = PM_PEER_ID_INVALID;
-static uint16_t m_conn_handle = BLE_CONN_HANDLE_INVALID;
-uint8_t passkey[6] = {1,2,3,4,5,6}; //FIXME must be random in the future
+uint16_t m_conn_handle = BLE_CONN_HANDLE_INVALID;
 //static ble_gap_sec_keyset_t keyset;
 // TODO add own UUIDs.
 static ble_uuid_t m_adv_uuids[] =                                                   /**< Universally unique service identifiers. */
@@ -174,11 +169,7 @@ void gap_params_init()
     err_code = sd_ble_gap_ppcp_set(&gap_conn_params);
     APP_ERROR_CHECK(err_code);
 
-    uint8_t passkey [] = "123456";
-    ble_opt_t optS;
-    optS.gap_opt.passkey.p_passkey=&passkey[0];
-    err_code = sd_ble_opt_set(BLE_GAP_OPT_PASSKEY, &optS);
-    APP_ERROR_CHECK(err_code);
+    init_passkey();
 }
 
 
@@ -387,27 +378,26 @@ static void ble_evt_handler(ble_evt_t const * p_ble_evt, void * p_context)
         break;
     case BLE_GAP_EVT_AUTH_KEY_REQUEST:
         NRF_LOG_INFO("BLE_GAP_EVT_AUTH_KEY_REQUEST received");
-        err_code = sd_ble_gap_auth_key_reply(
+        /*err_code = sd_ble_gap_auth_key_reply(
             conn_handle, 
             BLE_GAP_AUTH_KEY_TYPE_PASSKEY, 
             passkey);
-        APP_ERROR_CHECK(err_code);
+        APP_ERROR_CHECK(err_code);*/ //TODO not doing this anymore right?
         break;
     case BLE_GAP_EVT_PASSKEY_DISPLAY:
         {
-        //memcpy(p_ble_evt->evt.gap_evt.params.passkey_display.passkey, passkey, sizeof(passkey));
         char to_show[6 + 1];
         memcpy(to_show, p_ble_evt->evt.gap_evt.params.passkey_display.passkey, 6);
         to_show[6] = 0;
         NRF_LOG_INFO("Passkey: %s", nrf_log_push(to_show));
         } break;
-        case BLE_GAP_EVT_AUTH_STATUS:
-            NRF_LOG_INFO("BLE_GAP_EVT_AUTH_STATUS: status=0x%x bond=0x%x lv4: %d kdist_own:0x%x kdist_peer:0x%x",
-                p_ble_evt->evt.gap_evt.params.auth_status.auth_status,
-                p_ble_evt->evt.gap_evt.params.auth_status.bonded,
-                p_ble_evt->evt.gap_evt.params.auth_status.sm1_levels.lv4,
-                *((uint8_t *)&p_ble_evt->evt.gap_evt.params.auth_status.kdist_own),
-                *((uint8_t *)&p_ble_evt->evt.gap_evt.params.auth_status.kdist_peer));
+    case BLE_GAP_EVT_AUTH_STATUS:
+        NRF_LOG_INFO("BLE_GAP_EVT_AUTH_STATUS: status=0x%x bond=0x%x lv4: %d kdist_own:0x%x kdist_peer:0x%x",
+            p_ble_evt->evt.gap_evt.params.auth_status.auth_status,
+            p_ble_evt->evt.gap_evt.params.auth_status.bonded,
+            p_ble_evt->evt.gap_evt.params.auth_status.sm1_levels.lv4,
+            *((uint8_t *)&p_ble_evt->evt.gap_evt.params.auth_status.kdist_own),
+            *((uint8_t *)&p_ble_evt->evt.gap_evt.params.auth_status.kdist_peer));
         break;
 
     case BLE_GAP_EVT_LESC_DHKEY_REQUEST:
@@ -449,121 +439,6 @@ void ble_stack_init()
     NRF_SDH_BLE_OBSERVER(m_ble_obs, APP_BLE_OBSERVER_PRIO, ble_evt_handler, NULL);
 }
 
-
-/**@brief Function for handling events from the BSP module.
- *
- * @param[in]   event   Event generated by button press.
- */
-void bsp_event_handler(bsp_event_t event)
-{
-    ret_code_t err_code;
-
-    switch (event)
-    {
-        case BSP_EVENT_SLEEP:
-            sleep_mode_enter();
-            break;
-
-        case BSP_EVENT_DISCONNECT:
-            NRF_LOG_INFO("disconnecting due to BSP event");
-            err_code = sd_ble_gap_disconnect(m_conn_handle, BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
-            if (err_code != NRF_ERROR_INVALID_STATE)
-            {
-                APP_ERROR_CHECK(err_code);
-            }
-            break;
-
-        case BSP_EVENT_WHITELIST_OFF:
-            if (m_conn_handle == BLE_CONN_HANDLE_INVALID)
-            {
-                err_code = ble_advertising_restart_without_whitelist(&m_advertising);
-                if (err_code != NRF_ERROR_INVALID_STATE)
-                {
-                    APP_ERROR_CHECK(err_code);
-                }
-            }
-            break;
-
-        default:
-            break;
-    }
-}
-
-/**@brief Function for handling Peer Manager events.
- *
- * @param[in] p_evt  Peer Manager event.
- */
-static void pm_evt_handler(pm_evt_t const * p_evt)
-{
-    ret_code_t err_code;
-
-    pm_handler_on_pm_evt(p_evt);
-    pm_handler_disconnect_on_sec_failure(p_evt);
-    pm_handler_flash_clean(p_evt);
-
-    switch (p_evt->evt_id) {
-    case PM_EVT_CONN_SEC_SUCCEEDED:
-    {
-        // Check if the link is authenticated (meaning at least MITM).
-        pm_conn_sec_status_t conn_sec_status;
-        err_code = pm_conn_sec_status_get(p_evt->conn_handle, &conn_sec_status);
-        APP_ERROR_CHECK(err_code);
-
-        if (conn_sec_status.mitm_protected) {
-            NRF_LOG_INFO("Link secured. Role: %d. conn_handle: %d, Procedure: %d",
-                ble_conn_state_role(p_evt->conn_handle),
-                p_evt->conn_handle,
-                p_evt->params.conn_sec_succeeded.procedure);
-        } else {
-            // The peer did not use MITM, disconnect.
-            //NRF_LOG_INFO("Collector did not use MITM, disconnecting");
-            err_code = pm_peer_id_get(m_conn_handle, &m_peer_to_be_deleted);
-            APP_ERROR_CHECK(err_code);
-            err_code = sd_ble_gap_disconnect(m_conn_handle, BLE_HCI_REMOTE_USER_TERMINATED_CONNECTION);
-            APP_ERROR_CHECK(err_code);
-        }
-    } break;
-    case PM_EVT_CONN_SEC_FAILED:
-        m_conn_handle = BLE_CONN_HANDLE_INVALID;
-        break;
-    case PM_EVT_PEERS_DELETE_SUCCEEDED:
-        advertising_start(false);
-        break;
-    default:
-        break;
-    }
-}
-
-void peer_manager_init() {
-    uint32_t err_code;
-
-    NRF_LOG_INFO("before init");
-    err_code = pm_init();
-    APP_ERROR_CHECK(err_code);
-    NRF_LOG_INFO("after init");
-
-    // Security parameters to be used for all security procedures.
-    ble_gap_sec_params_t sec_param = {
-        .bond           = false,
-        .mitm           = true,
-        .lesc           = true,
-        .keypress       = false,
-        .io_caps        = BLE_GAP_IO_CAPS_DISPLAY_ONLY, // or BLE_GAP_IO_CAPS_KEYBOARD_ONLY
-        .oob            = false,
-        .min_key_size   = 7,
-        .max_key_size   = 16,
-        .kdist_own.enc  = false,
-        .kdist_own.id   = false,
-        .kdist_peer.enc = false,
-        .kdist_peer.id  = false,
-    };
-
-    err_code = pm_sec_params_set(&sec_param);
-    APP_ERROR_CHECK(err_code);
-
-    err_code = pm_register(pm_evt_handler);
-    APP_ERROR_CHECK(err_code);
-}
 
 /**@brief Function for initializing the Advertising functionality.
  */
