+++ src/pairing.h
@@ -1,0 +1,6 @@
+#include <stdint.h>
+
+void peer_manager_init();
+//void randomize_passkey();
+void update_passkey(uint8_t newkey[]);
+void init_passkey();
